Split private chat capability into a public chat capability and a private conversation capability (#1058)
* Split chat capability into a private and a public partition

The current (private) chat capability contains some functionality that
should be made public -- it was previously available using the
deeplink functionality. However, some should remain private since it
was previously private and will remain so pending future discussions.

* Delete previous files

* Update reference

* Change files

* Update changelog messagE

* Update test suite name
