+++ change/@microsoft-teams-js-56b791db-32a0-405b-88c5-e0deb0768544.json
@@ -1,0 +1,7 @@
+{
+  "type": "major",
+  "comment": "Split chat capability into a private (conversation) and a public (chat) partition",
+  "packageName": "@microsoft/teams-js",
+  "email": "aaroner@microsoft.com",
+  "dependentChangeType": "patch"
+}
